podcast_downloader/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

feedparser
azure-storage-blob
requests
python-dateutil==2.8.1
json